chore: release v0.2.1 with app icon and universal mac build
Adds a proper application icon (resources/icon.png) wired into both the
dev-mode dock and the packaged app via electron-builder. Switches the
macOS target from host-arch-only to a universal DMG/ZIP so Intel Macs
can install natively.

Also cleans up latent type errors: shared ambient types moved from
src/renderer/env.d.ts into src/shared/global.d.ts so main and preload
can see them, a course-manifest union-narrowing bug fixed, and a stale
tsconfig.web.json types reference corrected.
