+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "praxis",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "description": "Desktop course authoring and learning platform for technical developer training",
   "author": {
     "name": "Tim Pearson",
